requirements.txt
antlr4-python3-runtime==4.8
cycler==0.11.0
dataclasses==0.8
hydra-core==1.1.2
importlib-resources==5.2.3
joblib==1.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
omegaconf==2.1.2
opencv-python==4.5.5.62
Pillow==8.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
typing_extensions==4.0.1
zipp==3.6.0
